fix: keep the marker's fs graph out of the browser bundle
